proyects/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
cryptography
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
requests
python-multipart
pydantic
